extracted params from python file.
